fix(interpreter): fix compressed KDB deserialization
- Use static_cast<int> for ParseFromArray size parameter
- Remove debug output

+++ interpreter/src/kdb_builder.cpp
@@ -628,7 +628,7 @@ bool KdbBuilder::deserializeFromFileCompressed(const std::string& filepath) {
     }
     
     hwda::kdb::KnowledgeBase kdb;
-    if (!kdb.ParseFromArray(decompressed.data(), result)) {
+    if (!kdb.ParseFromArray(decompressed.data(), static_cast<int>(result))) {
         return false;
     }
